% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{checkAnalysed}
\alias{checkAnalysed}
\title{Check the analysis status
Function to determine if trial(s) has been analyzed - results sheet has any traits results}
\usage{
checkAnalysed(dir)
}
\arguments{
\item{dir}{String for the path to trial(s) files}
}
\description{
checkAnalysed
}
\keyword{analysis}
\keyword{trial,}
