% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{checkAnalyzed}
\alias{checkAnalyzed}
\title{Analyzed trials
Function to check if trial files are analyzed}
\usage{
checkAnalyzed(dir)
}
\arguments{
\item{dir}{for the directory containing the trial files}
}
\description{
checkAnalyzed
}
