% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{checkDataAvailable}
\alias{checkDataAvailable}
\title{Check data availability status of trials
A function to check if trial file(s) has any traits data in them}
\usage{
checkDataAvailable(dir)
}
\arguments{
\item{dir}{Path to the location of the trial file(s)}
}
\description{
checkDataAvailable
}
\keyword{data}
\keyword{trial,}
