% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{commonDir}
\alias{commonDir}
\title{The common location
A function to determine the common location of a list of paths (files, directories or both)}
\usage{
commonDir(nams)
}
\arguments{
\item{nams}{String list of paths}
}
\description{
commonDir
}
\keyword{location}
