% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{correctTrialNames}
\alias{correctTrialNames}
\title{correctTrialNames
 @title Use trial files to create trial name
Function that infers the names of the trial name based on the names of the trial files within the folder.}
\usage{
correctTrialNames(trialsRootFolder = NA)
}
\description{
correctTrialNames
 @title Use trial files to create trial name
Function that infers the names of the trial name based on the names of the trial files within the folder.
}
\keyword{breeder,}
\keyword{breeding}
\keyword{fieldbook,}
\keyword{inventories,}
\keyword{inventory,}
\keyword{nurseries,}
\keyword{nursery,}
\keyword{trial,}
\keyword{trials,}
