% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{fieldbookFile}
\alias{fieldbookFile}
\title{Identify the type of fieldbook file selected
A function to determine the fieldbook file type: stack, import, trial, etc}
\usage{
fieldbookFile(file)
}
\arguments{
\item{file}{Full path to an excel file}
}
\description{
fieldbookFile
}
\keyword{fieldbook,}
\keyword{file}
\keyword{import,}
\keyword{inventory,}
\keyword{nursery,}
\keyword{stack,}
\keyword{summary,}
\keyword{trial,}
