% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{getFirstPlot}
\alias{getFirstPlot}
\title{getFirstPlot
 @title Fieldbook first plot identification
Function to identify the location of the first plot in a sheet of a trial or nursery}
\usage{
getFirstPlot(file, sheet)
}
\arguments{
\item{file}{String for the path to the trial or nursery file}

\item{sheet}{String for the sheet of the file to check}
}
\description{
getFirstPlot
 @title Fieldbook first plot identification
Function to identify the location of the first plot in a sheet of a trial or nursery
}
\keyword{entry,}
\keyword{fieldbook,}
\keyword{inventories,}
\keyword{inventory,}
\keyword{nurseries,}
\keyword{nursery,}
\keyword{results}
\keyword{stock,}
\keyword{trial,}
\keyword{trials,}
