% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{hasStocks}
\alias{hasStocks}
\title{Status of stocklist sheet in a nursery file
A function to check if stockList has been generated}
\usage{
hasStocks(nursery_file)
}
\arguments{
\item{nursery_file}{Full path to the nursery file}
}
\description{
hasStocks
}
\keyword{entry}
\keyword{entrylist,}
\keyword{list,}
\keyword{nursery,}
\keyword{stock,}
\keyword{stocklist,}
