% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{importsPrepared}
\alias{importsPrepared}
\title{Compare trial files and respective import files
A function to check which files have respective import files}
\usage{
importsPrepared(dir)
}
\arguments{
\item{dir}{Path to directory containing the trial files}
}
\description{
importsPrepared
}
\keyword{import,}
\keyword{trial}
