% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCSV_functions.R
\name{ksuCSV}
\alias{ksuCSV}
\title{Prepare CSV files for trials for KSU use
Funtion to convert trial files to csv for KSU fieldbook field files}
\usage{
ksuCSV()
}
\description{
ksuCSV
}
\keyword{breeding,}
\keyword{collection,}
\keyword{data}
\keyword{fieldbook,}
\keyword{handheld}
\keyword{ksu,}
\keyword{trial,}
