% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{metadata}
\alias{metadata}
\title{Fieldbook data linkage and completeness function
Function to to connect data in different fieldbook files: inventory, nurseries and trials}
\usage{
metadata(inventories = NULL, nurseries = NULL, trials = NULL)
}
\arguments{
\item{inventories}{Directory path to the inventory file}

\item{nurseries}{Directory path to nursery files}

\item{trials}{Directory path to trial files}
}
\description{
metadata
}
\keyword{breeder,}
\keyword{breeding,}
\keyword{cgiar,}
\keyword{cimmyt,}
\keyword{fieldbook,}
\keyword{inventories,}
\keyword{inventory,}
\keyword{maize,}
\keyword{nurseries,}
\keyword{nursery,}
\keyword{stma}
\keyword{trial,}
\keyword{trials,}
