% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{Mode}
\alias{Mode}
\title{identify most common
A simple function to identify the most occurring category in data points}
\usage{
Mode(x)
}
\arguments{
\item{x}{A vector of datapoints}
}
\description{
Mode
}
\keyword{Mode}
