% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{removeBulking}
\alias{removeBulking}
\title{identify fixed line
A function to remove trailing `B` or `-` and `B*` to get a fixed line}
\usage{
removeBulking(x)
}
\arguments{
\item{x}{String for the pedigree or designation of the germplasm}
}
\description{
removeBulking
}
\keyword{fieldbook}
\keyword{germplasm,}
\keyword{inventory,}
\keyword{trial,}
