% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_functions.R
\name{separate_trials}
\alias{separate_trials}
\title{separate_trials
 @title Fieldbook files separation function
Function to check and separate different breeding files into the separate components:
inventories, nurseries, trials, seed preparations and stack files within the current directory.}
\usage{
separate_trials(dir = NA)
}
\description{
separate_trials
 @title Fieldbook files separation function
Function to check and separate different breeding files into the separate components:
inventories, nurseries, trials, seed preparations and stack files within the current directory.
}
\keyword{breeder,}
\keyword{breeding}
\keyword{fieldbook,}
\keyword{inventories,}
\keyword{inventory,}
\keyword{nurseries,}
\keyword{nursery,}
\keyword{trial,}
\keyword{trials,}
