% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateParent_function.R
\name{separateParent}
\alias{separateParent}
\title{utility function for separating the names for male and female from a pedigree name based on defined nomenclature}
\usage{
separateParent(pedigreeName)
}
\arguments{
\item{pedigreeName}{string for the pedigree name}
}
\description{
separateParent
}
\keyword{fieldbook,}
\keyword{inventory}
\keyword{nursery,}
\keyword{trial,}
