% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{splitString}
\alias{splitString}
\title{Split a string
A function to split a string using a given charater or string pattern}
\usage{
splitString(string, pattern, include = FALSE)
}
\arguments{
\item{string}{String to split}

\item{pattern}{Character or string name that is used to split the string}

\item{include}{Logical indicating whether to include include or remove the splitting character in the final split substrings}
}
\description{
splitString
}
