% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{stem}
\alias{stem}
\title{Common names
A function to identify the longest name that is common in a list of paths, files or names}
\usage{
stem(nams)
}
\arguments{
\item{nams}{The list of paths, files or names}
}
\description{
stem
}
\keyword{Stem}
