% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{typeBreeding}
\alias{typeBreeding}
\title{Check the level of
A function to extract the level of crosses in a nomenclature}
\usage{
typeBreeding(stringx)
}
\arguments{
\item{stringx}{A string for name or pedigree}
}
\description{
typeBreeding
}
\keyword{breeding,}
\keyword{nomenclature}
\keyword{pedigree,}
